Import into github
This is a script to filter out advert pages from pdf.

It is based on editorial content being coded text on the page whilst
advert is heavily based on graphic assets.
It can of course easily be defeated, so there's no warranty given.
The word count threshold may be adjusted.  This may actually become a
future option.

See README for usage.